post

+++ app/Http/Controllers/TasksController.php
@@ -1,3 +1,23 @@
+  public function index()
+    {
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $microposts = $user->microposts()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'microposts' => $microposts,
+            ];
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
+    }
+}
+
 <?php
 
 namespace App\Http\Controllers;
@@ -38,14 +58,22 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        // メッセージを作成
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->microposts()->create([
+            'content' => $request->content,
+        ]);
+        
+        // 前のURLへリダイレクトさせる
+        return back();
+        
+        /* メッセージを作成
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/'); */
     }
 
     // getでtasks/（任意のid）にアクセスされた場合の「取得表示処理」
@@ -96,12 +124,18 @@ class TasksController extends Controller
     // deleteでtasks/（任意のid）にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
+        }
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
     }
 }
